Update algorithmPolicy.py
Updated the algorithm to account for a potential workload policy file that can change certain workload values, account for a ct and tt file, account for special courses. Updated composite key, into a more dynamic group key that groups classes depending on more than just time and location as this caused issues due to inconsistencies in the data.